+++ packages/savefile/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Multiple include directories can now be given as a parameter
+
 ## [2.2.3]
 
 ### Fixed